Diffuse material
When each ray (light) hits the object, diffuse materials reflect the
ray in random directions. Since the reflected rays continue to move in
the scene, they also hit other objects, if any. This scenario is
similar to the movement of light in the real world. Therefore, a much
more realistic image is obtained.